# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: DraftFly API — AI-powered reply automation for B2B sales agencies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /personas:
    get:
      operationId: listPersonas
      tags: [personas]
      summary: List all personas
      parameters:
        - name: clientId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of personas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Persona"
    post:
      operationId: createPersona
      tags: [personas]
      summary: Create a persona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonaInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
  /personas/{id}:
    get:
      operationId: getPersona
      tags: [personas]
      summary: Get a persona
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Persona detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
        "404":
          description: Not found
    patch:
      operationId: updatePersona
      tags: [personas]
      summary: Update a persona
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PersonaUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
    delete:
      operationId: deletePersona
      tags: [personas]
      summary: Delete a persona
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ---------- CAMPAIGNS ----------
...
  /setup:
    get:
      operationId: listSetupItems
      tags: [setup]
      summary: Get setup checklist items
      parameters:
        - name: checklistType
          in: query
          schema:
            type: string
            enum: [client_onboarding, internal_setup]
      responses:
        "200":
          description: Setup items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SetupItem"
  /setup/{id}:
    patch:
      operationId: updateSetupItem
      tags: [setup]
      summary: Mark a setup item complete or incomplete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupItem"

  # ---------- DASHBOARD ----------
...
components:
  schemas:
...
    Persona:
      type: object
      required: [id, clientId, name, productDescription, targetAudience, toneOfVoice, cta, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        name:
          type: string
        productDescription:
          type: string
        targetAudience:
          type: string
        toneOfVoice:
          type: string
        commonObjections:
          type: ["string", "null"]
        cta:
          type: string
        qualificationRules:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    PersonaInput:
      type: object
      required: [clientId, name, productDescription, targetAudience, toneOfVoice, cta]
      properties:
        clientId:
          type: integer
        name:
          type: string
        productDescription:
          type: string
        targetAudience:
          type: string
        toneOfVoice:
          type: string
        commonObjections:
          type: string
        cta:
          type: string
        qualificationRules:
          type: string
    PersonaUpdate:
      type: object
      properties:
        name:
          type: string
        productDescription:
          type: string
        targetAudience:
          type: string
        toneOfVoice:
          type: string
        commonObjections:
          type: string
        cta:
          type: string
        qualificationRules:
          type: string

    # ---- Campaign ----
...
    SetupItem:
      type: object
      required: [id, title, description, category, checklistType, isCompleted]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [infrastructure, integrations, configuration, testing]
        checklistType:
          type: string
          enum: [client_onboarding, internal_setup]
        isCompleted:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
    SetupItemUpdate:
      type: object
      required: [isCompleted]
      properties:
        isCompleted:
          type: boolean

    # ---- Dashboard ----
...